+++ pangaea/datasets/xview2.py
@@ -163,7 +163,7 @@ class xView2(RawGeoFMDataset):
         img = torch.from_numpy(img.transpose((3, 0, 1, 2))).float()
         # img_pre = torch.from_numpy(img_pre.transpose((2, 0, 1))).float()
         # img_post = torch.from_numpy(img_post.transpose((2, 0, 1))).float()
-        msk = torch.from_numpy(msk).float()
+        msk = torch.from_numpy(msk).long()
 
 
         return {
